cmd/unity: pass workdir to testscript (#32)
Pre-calculating the workdir's for each of the script instances is
temporarily more expensive in terms of disk space (because we copy the
module N times for the N scripts, rather than doing it on demand) but
has the benefit that we can use git worktree add to do the "copy" rather
than our own copy implementation. This is always going to be more
correct.