cmd/unity: log results as a table (#49)
In either --corpus or project mode it is good to see the time it took
for each entry of the test matrix. Log that.

The first entry in a module's manifest's version list is used as the
base version. All subsequent tests of that module against different
version have their timings printed relative (as a % difference) to that
base version.