cmd/unity: fix Gerrit revision resolver test
With the switch to GerritHub, the Gerrit resolver resolves relative to a
different URL (GerritHub instead of GoogleSource). Hence the hard-coded
reference that we have as part of our test now fails when tested against
a commit of the CUE repo after (and including) 06484a39.

Fix this by providing a reference for a change in the GerritHub setup.

Signed-off-by: Paul Jolly <paul@myitcv.io>