cmd/unity: ensure cue is on PATH
testscript test commands that invoke cue currently work by virtue of a
testscript.Params.Cmds-defined command. If, however, such a command
itself invokes cue (possible via cue cmd) then this fails, because the
cue binary is not part of the PATH for the testscript evaluation.

Fix that, and add a test using cue cmd that verifies the behaviour.

Signed-off-by: Paul Jolly <paul@myitcv.io>