internal/ci: move workflows to their own package.
This helps prepare for a later change where we switch unity to having
GerritHub as a source of truth.

We also rename dispatch to unity, and test to trybot, again to minimise
later diffs.

This also copies the ci_tool.cue and vendor package from the core CUE
project which has progressed in the meantime.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I014767eeac0954a6a613b6b2f657422c6e9b8678