internal/ci: fix unity dispatch
The unity dispatch has cl and a versions fields in the payload, to
distinguish between what is a CL-based trigger (i.e. part of the
trybots) and a CLI-based trigger (i.e. someone testing from the command
line). A copy-paste-style error from the main CUE repo resulted in us
not dereferencing via the cl field for the changeID and commit. Fix
that.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I38dfd933c88a422579e17f97d8cabdc56136dc24