cmd/unity: add --skip-base flag to test command
We don't yet have caching of test results. As such, we always test the
base versions to verify a known good state. However, this can become
tiresome when testing a chain of CLs, especially where one of those CLs
is known to break unity. In that specific case, we want to use --update
in order to refresh our expectations, and then use that result for later
CLs in the chain. Not skipping base versions in this situation would
result in the next CL in the chain failing when it comes to running the
base versions (because expectations were updated from the previous CL
in the chain, the one with the breaking change).

With --skip-base, when we have such a breaking change CL, we can first
use --update, then git add, and then --skip-base with --staged on
subsequent CLs in the chain to verify that those subsequent CLs pass
with respect to the updated expectations.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I0fc9e213ed45b7bb5716f9a32f00c4852606fd84