cmd/unity: never reuse the current unity executable
I can see the intent behind it: if the binary exists,
and isn't a test binary with slightly different behavior,
and it already targets the platform we want, it's good enough.

Unfortunately, this is easily not true in multiple scenarios.
The one I ran into is that my host has glibc 2.35,
yet our current docker image appears to have a much older glibc,
resulting in failures to load glibc properly.
unity's build step already remedies this with CGO_ENABLED=0,
but since I was entering the "reuse the binary" happy path, it broke.

There are other possible scenarios where we want a new binary too,
such as if the current binary was built with different parameters:
an older version of Go, a different set of tags, no build/vcs info, etc.

To remove all these potential pitfalls, always build a binary.
The code already does its best to cache the built binary,
so this shouldn't have a big impact on runtime performance.

The only exception we make is when running in test scripts.
Since the parent test function already ran pathToSelf,
and the test script itself doesn't run inside the unity module,
it's both unnecessary and broken to try to build unity again.

While here, apply small code simplifications and add a TODO.

Change-Id: I87637eabf9235237f5bc1716845bba98d4df6575
Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Reviewed-on: https://review.gerrithub.io/c/cue-unity/unity/+/541991
TryBot-Result: <cue.porcuepine@gmail.com>
Reviewed-by: Paul Jolly <paul@myitcv.io>