support running Go tests in projects
A couple of major TODOs are left for follow-up CLs to properly complete
issue #41: obeying the Go and CUE versions when running the tests.

Some more TODOs are left for near-future improvements in this area.

We also add a testscript which exercises the basics.

Updates #41.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I909e547b0d69f3383ef37c6931752e425e8977bf
Reviewed-on: https://review.gerrithub.io/c/cue-unity/unity/+/543960
Reviewed-by: Paul Jolly <paul@myitcv.io>
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>