cmd/unity: fix with git 2.38.1 and fix typo
Due to CVE-2022-39253, git v2.38.1 flipped a config default to not allow
the "file" protocol for cloning from local directories.

We use it in our tests, and we're fine with that, so re-enable it.
Note that this is safe, as the "global" config in this context is inside
a temporary $HOME for each test script.

See: https://github.blog/2022-10-18-git-security-vulnerabilities-announced/#cve-2022-39253

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I725b8f8b47a90668bfcca60a272668e156e27eb1
Reviewed-on: https://review.gerrithub.io/c/cue-unity/unity/+/545219
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Paul Jolly <paul@myitcv.io>