cmd/unity: add an integration test for safe vs unsafe mode
While developing support for running Go tests,
I initially only implemented them in the unsafe mode.
A short while later I added support for the default safe mode,
by incorporating dockexec.

It was a surprise to me that the tests happily passed when the safe mode
was doing nothing in terms of containerization.
Add an integration test which will fail if a Unity test script or Go
test can peek into the host machine when they shouldn't (safe),
and fail if they aren't able when they should (unsafe).
See the added documentation in the test.

We add an "expand" builtin, which is similar to cmpenv, but replaces the
contents of a file with parameter expansion.
We need this to pass UNITY_HOST_HOME to the tests run by unity,
as only a few environment variables are passed through.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I94e414250a3d85e723507f02facce121778e5743