internal/ci: run Unity on CLs while comparing with the parent commit
As opposed to comparing with the base CUE version, which is often not
helpful for the wall time and other performance comparisons.

We use shell to fetch the Gerrit change and its parent commit,
since Unity expects both versions as arguments,
and we can't assume that the parent commit is in master.
Fetching both directly and checking them out in directories
is a straightforward way to pass the versions along to Unity.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I89d6e177d258b25810cfcb2a6f7fe2dbf7ccc13b
Reviewed-on: https://review.gerrithub.io/c/cue-unity/unity/+/549070
Reviewed-by: Paul Jolly <paul@myitcv.io>
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>