internal/ci: run Unity on CLs while comparing with the parent commit
As opposed to comparing with the base CUE version, which is often not
helpful for the wall time and other performance comparisons.

We use shell to fetch the Gerrit change and its parent commit,
since Unity expects both versions as arguments,
and we can't assume that the parent commit is in master.
Fetching both directly and checking them out in directories
is a straightforward way to pass the versions along to Unity.

This is the second attempt, as the first one was broken and reverted.
I mistakenly understood that GITHUB_REF was of the form
"unity/cl_number/patchset_number", when in fact it is like
"unity/change_id/git_commit".
Luckily, the commit is enough to perform a git fetch.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I03c1e09cad09684322f7129744e833548fc9e718
Reviewed-on: https://review.gerrithub.io/c/cue-unity/unity/+/549178
Reviewed-by: Paul Jolly <paul@myitcv.io>
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>