internal/ci: try to fix trybot and unity triggers again
In the previous CL, we updated internal/ci/gerrithub so that it would
use the five-part branch name to include the CL and patchset numbers.

However, I missed that unity_dispatch.yml does not actually use that
gerrithub package to template the dispatch workflow.
Instead, a subset of what the gerrithub package does is copy-pasted.

Following the same trend for now, copy-paste the extra bit which sets up
those two more branch name parts.
The bits are copy-pasted verbatim from the gerrithub package,
but using `payload.cl.ref` rather than `payload.ref`.

We also missed another crucial aspect of how CI on cue-unity/unity
differs from cue-lang/cue. Namely, we don't use PRs to run Unity;
cue-lang/cue started doing that recently to better use actions/cache.
We don't need to do that here, nor do we have a token set up with a bot
account to perform `gh auth login`, so avoid that entirely.
Trying to do that login without a token was hanging the job on a prompt.

Since we're not using PRs to trigger the actual Unity runs,
we need to tweak the branch name pattern to use five parts,
mirroring the other recent changes to drop the three-part branch names.
The glob-like asterisk does not match slashes, crucially.

Finally, the workflow which does run Unity was fetching the commit hash
from $GITHUB_REF via basename, which was OK when the commit was the last
path-like part of the branch. It is not anymore; update the script.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I9b9d4f261a02e94c4b6e783be8d589d0df101e7c