internal/ci: checkout submodules for unity CLI run.
We were not checking out submodules as part of a unity CLI run. Hence we
were never doing any work testing against the corpus.

The reason this was never caught is that unity is silent when it tests
nothing in corpus mode. This is arguably an unsafe default behaviour,
which we fix in the next CL.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I2f0cf4bb2b3198146f20eaa7ef52821a884abf41