internal/ci: establish repo package
This establishes a repo package in line with the main CUE repo. We move
some common declarations from workflows.cue to this new package.

Also move away from using definitions in the new repo package: it
contains concrete configuration values.

Also import the repo package once in the github package and expose as
_repo, again following the pattern of the main CUE repo.

This change should not result in any .github/workflows changes.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I3c71ec665c11b005e819ba2ed6b9ea946021ff8b