internal/ci: baseline base against ABCDEF from main CUE repo
DO NOT SUBMIT.

Need to update this commit message with the actual commit we baseline
against.

The contents of the base package are a mechanical copy from be0601bf in
the main CUE repo.

The changes to all other internal/ci/**/*.cue packages are mechanical
changes in response to that new baseline.

The changes in .github/workflows should strictly improve
correctness/reliability. The net effect of the changes however is
expected to be zero, i.e. site will still deploy a preview, master will
continue to deploy etc.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I0139482e75cdae3873446818e152967974d18719