internal/ci: baseline base against f881884a from main CUE repo
The contents of the base package are a mechanical copy from f881884a in
the main CUE repo.

We switch all workflows in this repo to be linux-only, moving away from
the matrix pattern which is only required if we are running on multiple
Go versions or runners.

We also adopt the base.trybotWorkflows template to DRY up our github
package. This means we can drop the manually maintained trybot_dispatch
file, but that we gain automatically templated versions of
trybot_dispatch, push_tip_to_trybot and evict_caches.

Changes to the trybot workflow are strict improvements in reusing base
functionality (caching, git file modtime restore, go caching).

We drop the gerrithub package because that functionality is part of the
mechanically copied base.

We add to steps.cue with some steps common to various workflows.

We drop use of isLatestLinux entirely because we no longer use a build
matrix for the trybot workflow.

We ensure the repo package embeds base, per the other repos.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I0139482e75cdae3873446818e152967974d18719