internal/ci: baseline against main CUE repo post CI changes
This needs to wait until CL 551352 lands.

The medium-long term intention is that the base package should be
extracted to a module shared in some way by all repos.

For now we copy-paste from the main repo where the recent changes have
been made.

This brings the master branch of this repo in line with CL 551380 in the
main repo.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I21ff6072944d37f9551c656253ef0f530e9c1a3f